### 2018-10-09 #997 @ngoldbaum
Thanks for the ping, I'd love to do this but unfortunately I think I can't review since I'm also employed by NCSA, at least according to [the COI policy](https://github.com/openjournals/joss/blob/master/COI.md).
