### 2018-10-09 #997 @ygrange
Looks really interesting and relevant. I have to admit though that available time may be an issue for me.
