### 2018-10-09 #997 @arfon
> Thanks for the ping, I'd love to do this but unfortunately I think I can't review since I'm also employed by NCSA, at least according to the COI policy.

Do you feel able to give an independent review of this software? If so, we can give a waiver in this case.
