### 2018-10-11 #997 @wtgee
Oops, sorry for late reply. Probably too busy for this right now but can
make time if Nathan is unavailable.

Cheers!

On Wed, Oct 10, 2018 at 5:58 AM Arfon Smith <notifications@github.com>
wrote:

> Thanks for the ping, I'd love to do this but unfortunately I think I can't
> review since I'm also employed by NCSA, at least according to the COI
> policy.
>
> Do you feel able to give an independent review of this software? If so, we
> can give a waiver in this case.
>
