### 2018-10-14 #997 @arfon
> Thanks for the ping, I'd love to do this but unfortunately I think I can't review since I'm also employed by NCSA, at least according to the COI policy.

@ngoldbaum - just following up on my [comment above](https://github.com/openjournals/joss-reviews/issues/997#issuecomment-428310389). I'm happy to have you review if you feel like you can give an impartial review.
