### 2018-10-15 #997 @arfon
@ngoldbaum @mgckind - see you both over in #1022 for the actual review.
