### 2018-10-29 #1022 @mgckind
Hi @ngoldbaum, thanks again for reviewing this, let me know if you'd need anything from our side including maybe a special access to DES account.
