### 2018-11-05 #1022 @ngoldbaum
Comments on the paper text:

The JOSS reviewer guidelines note:

> Note the paper should not include software documentation such as API (Application Programming Interface) functionality, as this should be outlined in the software documentation.

So I think Figure 3 and the "architecture" section probably belong in the docs rather than in the paper.

I also think "installation" probably also belongs in the docs. I think it would be sufficient to say "API documentation and installation instructions are available on the online documentation, which you can link to in the text.

Language nits:

* In the summary, the phrase "and support customized" should be "and supports customized"

* In the second setence of the second paragraph, "goals of DES are reveal" should be "goals of DES are to reveal".
