### 2018-11-05 mgckind/easyaccess#158 @ngoldbaum
More graceful failure when oracle client library is not installed
Current if the oracle client library can't be located (for example, if someone installs `easyaccess` from pip but fails to install the oracle binaries), this happens:

```
$ easyaccess -s desdr

Error in DES_SERVICES config file, creating a new one...
File might not exists or is not configured correctly

Adding section db-desdr to des_service file

Enter username : ngoldbaum
Enter password :

DPI-1047: 64-bit Oracle Client library cannot be loaded: "dlopen(libclntsh.dylib, 1): image not found". See https://oracle.github.io/odpi/doc/installation.html#macos for help
```

While the error message does point to the installation instructions for the oracle client, this is such a common pitfall that the error message itself should try to explain what's going wrong in less technical language, something like "easyaccess was unable to load the oracle client library, please follow the installation instructions below:"

I would also encourage you to more prominently explain in the installation instructions in the readme that the oracle client library needs to be installed separately. I'm not sure offhand if that applies only to the pip installation (what I used) or to the conda installation as well.

See https://github.com/openjournals/joss-reviews/issues/1022
