### 2018-11-05 #1022 @ngoldbaum
I'm not a lawyer and haven't done much more than cursory googling on the subject, but I'm a little concerned about the direct binary dependency on the oracle client library, which is licensed under then OTN license. Among the terms are the following text:

> If You want to use the Programs or Your application for any purpose other than as expressly permitted under this Agreement, You must obtain from Oracle or an Oracle reseller a valid Programs license under a separate agreement permitting such use. However, You acknowledge that the Programs may not be intended for production use and/or Oracle may not make a version of the Programs available for production or other purposes; any development or other work You undertake with the Programs is at Your sole risk.

I read this to say that you're only allowed to write code that depends on OTN-licensed software for testing and development purposes, but must get an oracle license or permission from oracle for production work.
