### 2018-11-05 mgckind/easyaccess#159 @ngoldbaum
Oracle client install instructions in the readme are not publicly visible
The readme links to a confluence page for instructions on how to properly set up the oracle client library, which is not publicly visible:

https://opensource.ncsa.illinois.edu/confluence/display/DESDM/Instructions+for+installing+Oracle+client+and+easyaccess+without+EUPS

The contents of this wiki page should be part of the easyaccess docs, since easyaccess seems to crash when I try to login to the database without the client library installed, so it's very important to make it easy and clear how to install that dependency.

It was also not clear to me until I tried to install the client libraries that I'd need to sign up for an oracle account to even get to the download page.

see https://github.com/openjournals/joss-reviews/issues/1022
