### 2018-11-05 #1022 @ngoldbaum
Regarding the previous comment, I think this means that people can only use `easyaccess` for production work on a system where someone has paid for an oracle license (e.g. on an NCSA system), but if someone installs `easyaccess` on another system or e.g. their laptop, then they *must not* use `easyaccess` for production work without risking litigation from oracle.

I'm not sure if there's a license expert associated with JOSS who can weigh in here @arfon.
