### 2018-11-05 #1022 @arfon
> I'm not sure if there's a license expert associated with JOSS who can weigh in here @arfon.

Yeah, I'm really not sure here. On the one hand, I guess this statement may just reflect the licensing terms of Oracle, i.e. you can't use an Oracle DB server in production without a license. That said, these terms do seem to run somewhat counter-culture to the 'Free' in FLOSS.

@Kevin-Mattheus-Moerman - perhaps we can get some help from the OSI on this topic? Would you be able to seek their assistance on our behalf.

Also, @mgckind - is this something you've seen discussed before at NCSA?
