### 2018-11-05 #1022 @mgckind
Thanks, `easyaccess`  as `cx_Oracle` and others are just clients talking with the Oracle DB installed, licensed and running at NCSA, you'd only need the instant client library from Oracle, which you can obtain from their page.
we talked about this in the past and was suggested (it's been a while) that given the nature of the closed collaboration that we had on issues, for the public facing part we might recommend to get the Oracle client from them before installing easyaccess, which in addition (in the public case) is a read-only system just to extract data.
In any case installing these tools doesn't allow to run your own Oracle DB server, just the client.
