### 2018-11-05 #1022 @ngoldbaum
Right, the issue is that the client itself is distributed under the OTN license. Since `easyaccess` depends on it, if I have `easyaccess` installed on a system where there isn't an oracle license available and I use it for production work, then I'm breaking the Oracle client license. The relevant part of the license is this bit (emphasis mine):

> If You want to use the Programs *or Your application* for any purpose other than as expressly permitted under this Agreement

Here "the Programs" is the oracle client library and "your application" is `easyaccess`.

By the way, I don't think this issue should prevent getting an `easyaccess` paper published in JOSS, however we should definitely figure out if this is a real problem for potential users and if so make it extremely clear in the readme and documentation what the legal issues are.
