### 2018-11-05 mgckind/easyaccess#159 @ngoldbaum
Just a heads up that I'm blocked on going further on my peer review by this one as I'm not sure how to install the client library to allow me to use easyaccess locally.
