### 2018-11-07 mgckind/easyaccess#158 @mgckind
Thanks, It's true. That error comes directly from Oracle and can't be modified, we can add a test to check whether the library was properly installed (given that cx_Oracle is a requirement, it checks for the installation of the Oracle client first, but not whether is properly installed)
