### 2018-11-07 mgckind/easyaccess#159 @mgckind
Thanks, The main purpose of the library is for the DES collaboration and those are instructions are available for them.
Can you try using conda?
I can also provide you with access to the internal DB to look at some of the functionalities.
Thanks for doing the review!
