### 2018-11-07 #1022 @mgckind
Thanks again,
I went and look again and I think by being more clear in the requisites and instructions for installation we can solve this. `easyaccess` requires `cx_Oracle` which is a python wrapper for the Oracle client which is also a requirement. In the `cx_Oracle` [page](https://cx-oracle.readthedocs.io/en/latest/installation.html#installing-cx-oracle-on-linux) they describe the installation of the oracle client as a requirement and they have an open-source license for the wrapper.
In short, we just need to be more clear about meeting the requirements in the installation process, but there won't be any issues using `easyaccess`
