### 2018-11-07 mgckind/easyaccess#158 @ngoldbaum
Yeah, I'd suggest adding import guards around the `cx_Oracle` import, capture the exception, and then reraise with a better error message.
