### 2018-11-12 mgckind/easyaccess#159 @mgckind
Were you able installing using conda? I have  a test account setup for you to log in into DES.
