### 2018-11-12 mgckind/easyaccess#158 @mgckind
The problem with that is that every Oracle Error is different, from syntax errors to connection errors, locks, etc. I have those guards and actually print the errors instead of raising an Exception. I can definitely add something regarding the first time running to make sure everything is ok.
