### 2018-11-13 mgckind/easyaccess#158 @ngoldbaum
Makes sense, yes at least a mention in the readme would be good. It definitely confused me :)
