### 2018-11-13 mgckind/easyaccess#159 @ngoldbaum
You don't think it would be possible to expose the installation instructions in a publicly visible location? I understand that this will be mostly used by the DES collaboration, but I don't think that should prevent people from setting things up on their own system if that's possible, especially if the only thing blocking that is making some documentation that already exists publicly viewable.

I haven't tried to set this up yet using conda, I've been working on other things. My plan was to come back to this today.
