### 2018-11-13 #1022 @ngoldbaum
Hmm @arfon am I supposed to actually click the check marks in the comment from whedon at the top? I don't seem to be able to click the check boxes.
