### 2018-11-13 mgckind/easyaccess#160 @ngoldbaum
Conda installation does not include pandas
As outlined in the README, I did:

```
conda install easyaccess==1.4.5 -c mgckind
```

This installed the following list of packages on top of a bare miniconda installation:

```
The following NEW packages will be INSTALLED:

    blas:                 1.0-mkl
    cx_oracle:            5.3-py36_0             mgckind
    easyaccess:           1.4.5-py36h079ba9d_0   mgckind
    fitsio:               0.9.11-py36_1          mgckind
    future:               0.17.1-py36_0
    intel-openmp:         2019.0-118
    libgfortran:          3.0.1-h93005f0_2
    mkl:                  2019.0-118
    mkl_fft:              1.0.6-py36hb8a8100_0
    mkl_random:           1.0.1-py36h5d10147_1
    numpy:                1.15.4-py36h6a91979_0
    numpy-base:           1.15.4-py36h8a80b8c_0
    oracle-instantclient: 11.2.0.4.0-2           mgckind
    python-dateutil:      2.7.5-py36_0
    termcolor:            1.1.0-py36_0           mgckind
```

This doesn't include pandas, which causes an issue later when I try to use easyaccess:

```
$ easyaccess
Traceback (most recent call last):
  File "/Users/goldbaum/miniconda3/bin/easyaccess", line 2, in <module>
    import easyaccess as ea
  File "/Users/goldbaum/miniconda3/lib/python3.6/site-packages/easyaccess/__init__.py", line 1, in <module>
    from .easyaccess import *
  File "/Users/goldbaum/miniconda3/lib/python3.6/site-packages/easyaccess/easyaccess.py", line 18, in <module>
    import easyaccess.eautils.fileio as eafile
  File "/Users/goldbaum/miniconda3/lib/python3.6/site-packages/easyaccess/eautils/fileio.py", line 14, in <module>
    import pandas as pd
ModuleNotFoundError: No module named 'pandas'
```

Presumably the conda recipe you're using to build packages for your channel don't list pandas as a runtime dependency.
