### 2018-11-13 mgckind/easyaccess#161 @ngoldbaum
Unable to access DES data
Not sure if this is the correct place, but I guess I need to take you up on your offer to give me database access manually. I tried filling out the form linked in the readme here: https://des.ncsa.illinois.edu/easyweb/signup/

That did work, but I'm unable to use this login to actually access the database. An account under my e-mail (ngoldbau@illinois.edu) does seem to exist though and I'm able to reset my password, I'm just not able to actually use the account I've made to login.
