### 2018-11-13 mgckind/easyaccess#162 @ngoldbaum
Document license issues when using oracle instant client
Easyaccess depends on the oracle client library. As raised in the JOSS review (https://github.com/openjournals/joss-reviews/issues/1022), this might lead to potential legal issues if someones makes use of easyaccess for production work on a system that hasn't been licensed for production work with Oracle.

First, we need to verify that this is a real issue. My reading of the OTN license is that yes this is an issue but I'm not a lawyer. @arfon pinged @Kevin-Mattheus-Moerman in the JOSS review issue, perhaps he can weigh in?

If it is a real problem then this concern needs to be explicitly mentioned in the documentation to warn users about potential legal issues when they use easyaccess for production work.
