### 2018-11-13 mgckind/easyaccess#163 @ngoldbaum
Add contribution, support instructions
Currently I don't think there are any explicit instructions for how to contribute to `easyaccess` development or how to report issues. Please add a brief mention in the readme and possibly also a more detailed CONTRIBUTING file.

Once this is done I'll check the "community guidelines" checkbox in the JOSS review checklist.

See openjournals/joss-reviews#1022
