### 2018-11-13 mgckind/easyaccess#161 @mgckind
I'll send you a DES account internally by email, also for the public data you migth need to use `easyaccess -s desdr`
