### 2018-11-14 mgckind/easyaccess#162 @ngoldbaum
Another issue - are there license issues with distributing the oracle instant client on your conda channel? When I downloaded it from oracle I needed to sign up for an oracle account and explicitly agree to the OTN license before downloading the code.

In particular, in the OTN license it says that redistributors must (among other things):

> (b) keep executed end user agreements and records of end user information including name, address, date of distribution and identity of Programs distributed; (c) allow us to inspect Your end user agreements and records upon request; and, (d) enforce the terms of Your end user agreements so as to effect a timely cure of any end user breach, and to notify us of any breach of the terms.

Just running a conda channel doesn't do that.
