### 2018-11-15 mgckind/easyaccess#164 @ngoldbaum
Add FAQ section to the readme
Matias mentioned to me over private e-mail that there's a FAQ page on a DES confluence about easyaccess. I'd urge you to move or duplicate that information in the readme for easyacess for the benefit of any users outside the DES collaboration.

In particular, there is an entry mentioning to edit the /etc/hosts file to fix an authentication issue that I ran into.

openjournals/joss-reviews#1022
