### 2018-11-15 mgckind/easyaccess#161 @ngoldbaum
I was able to clear this up, closing
