### 2018-11-15 mgckind/easyaccess#165 @ngoldbaum
Consider creating narrative documentation for easyaccess
Some of the usage example movies in the presentation describing easyaccess are a little hard to follow. In addition, it's a little awkward to search a presentation for information and probably also not straightfoward to add new content to.

You may want to consider creating a set of sphinx documentation for easyaccess, which you could then deploy for free on readthedocs or github pages. This will produce a set of HTML documentation that might be a bit easier for your users to interact with. It would also be a good place to migrate content from the DES confluence to a public space if you care at all about that.
