### 2018-11-15 #1022 @ngoldbaum
Matias has kindly sent me a login and I was able to access the DES database via easyaccess. I was also able to run some of the examples in the presentation on the easyacess website.

Regarding installation and initial configuration issues I ran into I've opened the following issues:

- [ ] https://github.com/mgckind/easyaccess/issues/158
- [ ] https://github.com/mgckind/easyaccess/issues/159
- [ ] https://github.com/mgckind/easyaccess/issues/160
- [ ] https://github.com/mgckind/easyaccess/issues/164

Regarding license issues and documentation of legal issues around the oracle client library:

- [ ] https://github.com/mgckind/easyaccess/issues/162

Regarding documentation on contribution, bug reports and questions about the software:

- [ ] https://github.com/mgckind/easyaccess/issues/163

I've also separately opened https://github.com/mgckind/easyaccess/issues/165, which is more of a "nice to have" rather than something I'm insisting on as part of the peer review process.

As soon as The issues I've added tasks for above are resolved I will go ahead and resolve the remaining check marks in the review checklist and ask @arfon to more forward with publishing the paper.

Overall the experience using easyacess was very pleasant and the examples in the reveal presentation on the website are very clear. I can see why this tool is so popular inside the DES collaboration. Hopefully publishing the paper about it will allow you to get more credit from people who use it as part of their daily workflow.
