### 2018-11-15 mgckind/easyaccess#165 @mgckind
Thanks, it's been on my TODO for some time, the presentation grew to be more than I originally planned
