### 2018-12-07 mgckind/easyaccess#162 @arfon
Thanks for opening this @ngoldbaum - I've asked NumFOCUS legal to help us resolve this question.
