+++ easyaccess/eaparser.py
@@ -8,9 +8,9 @@ import os
 
 class MyParser(argparse.ArgumentParser):
     def error(self, message):
-        print('\n*****************')
-        sys.stderr.write('error: %s \n' % message)
-        print('*****************\n')
+        print("\n*****************")
+        sys.stderr.write("error: %s \n" % message)
+        print("*****************\n")
         self.print_help()
         sys.exit(2)
 
@@ -18,48 +18,94 @@ class MyParser(argparse.ArgumentParser):
 def get_args(config_file):
     conf = config_mod.get_config(config_file)
     parser = MyParser(
-        description='Easy access to the DES database. There is a configuration file '
-                    'located in %s for more customizable options' % config_file)
-    parser.add_argument("-v", "--version", action="store_true",
-                        help="print version number and exit")
-    parser.add_argument("-c", "--command", dest='command',
-                        help="Executes command and exit")
-    parser.add_argument("-l", "--loadsql", dest='loadsql',
-                        help="Loads a sql command, execute it and exit")
-    parser.add_argument("-lt", "--load_table", dest='loadtable',
-                        help="Loads data from a csv, tab, or fits formatted file \
+        description="Easy access to the DES database. There is a configuration file "
+        "located in %s for more customizable options" % config_file
+    )
+    parser.add_argument(
+        "-v", "--version", action="store_true", help="print version number and exit"
+    )
+    parser.add_argument(
+        "-c", "--command", dest="command", help="Executes command and exit"
+    )
+    parser.add_argument(
+        "-l",
+        "--loadsql",
+        dest="loadsql",
+        help="Loads a sql command, execute it and exit",
+    )
+    parser.add_argument(
+        "-lt",
+        "--load_table",
+        dest="loadtable",
+        help="Loads data from a csv, tab, or fits formatted file \
                         into a DB table using the filename as the table name or a custom \
-                        name with --tablename MYTABLE. Not available in desdr.")
-    parser.add_argument("-at", "--append_table", dest='appendtable',
-                        help="Appends data from a csv, tab, or fits formatted file \
+                        name with --tablename MYTABLE. Not available in desdr.",
+    )
+    parser.add_argument(
+        "-at",
+        "--append_table",
+        dest="appendtable",
+        help="Appends data from a csv, tab, or fits formatted file \
                         into a DB table using the filename as the table name or a custom \
-                        name with --tablename MYABLE. Not available in desdr.")
-    parser.add_argument("--tablename", dest='tablename',
-                        help="Custom table name to be used with --load_table\
-                        or --append_table. Not available in desdr.")
-    parser.add_argument("--chunksize", dest='chunksize', type=int, default=None,
-                        help="Number of rows to be inserted at a time. Useful for large files "
-                             "that do not fit in memory. Use with --load_table or --append_table")
-    parser.add_argument("--memsize", dest='memsize', type=int, default=None,
-                        help=" Size of chunk to be read at a time in Mb. Use with --load_table or "
-                             "--append_table")
-    parser.add_argument("-s", "--db", dest='db',
-                        choices=['dessci', 'desoper', 'destest', 'desdr'],
-                        help="Override database name [dessci,desoper,destest,desdr]")
-    parser.add_argument("-q", "--quiet", action="store_true", dest='quiet',
-                        help="Silence initialization, no loading bar")
-    parser.add_argument("-u", "--user", dest='user')
-    parser.add_argument("-p", "--password", dest='password')
-    parser.add_argument("-nr", "--no_refresh", dest='norefresh', action="store_true",
-                        help="Do not refresh metadata at starting up to speed initialization. "
-                             "Metadata can always be refreshed from inside using the "
-                             "refresh_metadata command")
[…]
