+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changes
 
+## v1.4.6
+#### 2019-JAN-00
+-
+-
+-
+
+
 ## v1.4.5
 #### 2018-OCT-04
 - Restructuring code structure (PR #151)
