### 2019-01-11 8dc12d1
Add contribution and code of conduct
