### 2019-01-11 bc155cf
Add contribution and code of conduct
