### 2019-01-11 260268f
update conda config files
