+++ config/conda/meta.yaml
@@ -20,6 +20,7 @@ requirements:
     - python
     - future >=0.15.0
     - requests
+    - pandas
     - termcolor >=1.1.0
     - fitsio >=0.9.11
     - oracle-instantclient ==11.2.0.4.0
@@ -31,6 +32,7 @@ requirements:
     - python
     - future >=0.15.0
     - requests
+    - pandas
     - termcolor >=1.1.0
     - fitsio >=0.9.11
     - oracle-instantclient ==11.2.0.4.0
@@ -41,7 +43,7 @@ requirements:
 test:
   # Python imports
   imports:
-    #- easyaccess
+    - easyaccess
     - cx_Oracle
     - fitsio
 
