### 2019-01-11 mgckind/easyaccess#160 @mgckind
Now it does, I've added the missing dependencies, built new packages and replace the files for the current version (1.4.5), previous versions didn't have this issue and it will be added for the next 1.4.6 version in preparation. Thanks!
