### 2019-01-11 mgckind/easyaccess#163 @mgckind
I've added a contributions guideline, code of conduct and author files.  They are currently in the [1.4.6 branch](https://github.com/mgckind/easyaccess/tree/1.4.6) soon to be merged. Thanks
