### 2019-01-11 mgckind/easyaccess#160 @mgckind
I've updated the Docker images as well. Sorry this took so long...
