### 2019-01-14 #1022 @mgckind
Yes, @arfon. As a matter of fact, I started to work on the issues raised by @ngoldbaum last week and I should be able to finish them pretty soon, thanks!
