+++ README.md
@@ -1,15 +1,16 @@
-# easyaccess <a href="https://github.com/mgckind/easyaccess/releases/tag/1.4.6"> <img src="https://img.shields.io/badge/release-v1.4.6-blue.svg" alt="latest release" /></a> <a href="https://github.com/mgckind/easyaccess/blob/master/LICENSE.txt"><img src="https://img.shields.io/badge/license-NCSA%20License-blue.svg" alt="License" /> </a> <a href="https://pypi.python.org/pypi/easyaccess/1.4.6"><img src="https://img.shields.io/badge/pypi-v1.4.6-orange.svg" alt="pypi version"/></a>![](https://img.shields.io/conda/v/mgckind/easyaccess.svg)
+# easyaccess <a href="https://github.com/mgckind/easyaccess/releases/tag/1.4.6"> <img src="https://img.shields.io/badge/release-v1.4.6-blue.svg" alt="latest release" /></a> <a href="https://github.com/mgckind/easyaccess/blob/master/LICENSE.txt"><img src="https://img.shields.io/badge/license-NCSA%20License-blue.svg" alt="License" /> </a> <a href="https://pypi.python.org/pypi/easyaccess/1.4.6"><img src="https://img.shields.io/badge/pypi-v1.4.6-orange.svg" alt="pypi version"/></a> ![](https://img.shields.io/conda/v/mgckind/easyaccess.svg) ![](https://img.shields.io/conda/pn/mgckind/easyaccess.svg) ![](https://img.shields.io/badge/python-2.7%7C3.6-blue.svg)
+
+Enhanced command line SQL interpreter client for astronomical surveys.
 ![help_screen](data/help.gif)
 
-Enhanced command line SQL interpreter client for astronomical databases.
-
-Python Command Line Interpreter to access Oracle DES DB
-using cx_Oracle
+## Description
+`easyaccess` is an enhanced command line interpreter and Python package created to facilitate access to astronomical catalogs stored in SQL Databases. It provides a custom interface with custom commands and was specifically designed to access data from the Dark Energy Survey Oracle database, including autocompletion of tables, columns, users and commands, simple ways to upload and download tables using csv, fits and HDF5 formats, iterators, search and description of tables among others. It can easily be extended to another surveys or SQL databases. The package was completely written in Python and support customized addition of commands and functionalities.
 
 For a short tutorial check [here](http://matias-ck.com/easyaccess)
 
-**Current version = 1.4.5**
+**Current version = 1.4.6**
 
+#### DES DR1 users
 For DES public data release, you can start `easyaccess` with:
 
     easyaccess -s desdr
@@ -18,24 +19,89 @@ To create an account click [here](https://des.ncsa.illinois.edu/easyweb/signup/)
 
 ## Requirements
 
-- Oracle Client > 11g.2 (External library, no python)
-  Check [here](https://opensource.ncsa.illinois.edu/confluence/display/DESDM/Instructions+for+installing+Oracle+client+and+easyaccess+without+EUPS) for instructions on how to install these libraries
-- [cx_Oracle](https://bitbucket.org/anthony_tuininga/cx_oracle)
-
-  Note that cx_Oracle needs libaio on some Linux systems (e.g., #98)
-
-  Note that cx_Oracle needs libbz2 on some Linux systems
+- [Oracle Client](https://www.oracle.com/technetwork/database/database-technologies/instant-client/overview/index.html) > 11g.2 (External library, no python)
+  Check [here](https://www.oracle.com/technetwork/database/database-technologies/instant-client/overview/index.html) for instructions on how to install these libraries
+- [cx_Oracle](https://cx-oracle.readthedocs.io/en/latest/index.html)
+  - Note that cx_Oracle needs libaio on some Linux systems
+  - Note that cx_Oracle needs libbz2 on some Linux systems
[…]
