### 2019-01-14 2da7ac2
update paper in regards https://github.com/openjournals/joss-reviews/issues/1022
