### 2019-01-14 mgckind/easyaccess#158 @mgckind
I've updated the Readme with clear instructions and also added a FAQ
