### 2019-01-14 mgckind/easyaccess#159 @mgckind
I've added clear instructions and remove the links to internal pages, also found out
the  oracle client is indeed free for distribution
https://www.oracle.com/technetwork/database/database-technologies/instant-client/overview/index.html
