### 2019-01-14 mgckind/easyaccess#162 @mgckind
The Oracle instant client which is required by `cx_Oracle` and then by `easyaccess` link is here:https://www.oracle.com/technetwork/database/database-technologies/instant-client/overview/index.html I think there is no issues here since we are nos distributing Oracle by any means, just the client to access the DBs. I've added extra text in the Readme on how to install the client.
https://www.oracle.com/technetwork/database/database-technologies/instant-client/overview/index.html
