### 2019-01-14 mgckind/easyaccess#165 @mgckind
I've going to update the current page, but want to leave this open to eventually also move to readthedocs including the docstrings from the methods inside the code, Thanks!
