+++ config/conda/meta.yaml
@@ -1,9 +1,9 @@
 package:
   name: easyaccess
-  version: "1.4.5"
+  version: "1.4.6"
 
 source:
-  git_rev: 1.4.5
+  git_rev: 1.4.6
   git_url: https://github.com/mgckind/easyaccess.git
 
 build:
