### 2019-01-14 #1022 @mgckind
Hi, @arfon and @ngoldbaum.
I've addressed the issues mentioned above, update the paper also according to @ngoldbaum  comments and it's ready for revision. I've created a new version (1.4.6) but just to reflect the changes in the documentation, no changes in the code.
In regards Oracle,
I mentioned in the issues, but also I'll repeat it here:
`easyaccess` uses the python Oracle client `cx_Oracle` ([here](https://cx-oracle.readthedocs.io/en/latest/index.html)) which is from Oracle, they have instructions on how to install it. An external requirement for this wrapper is to have the Oracle Instant Client installed (just the client, not the server), which is "Free to download, deploy and distribute" see [here](https://www.oracle.com/technetwork/database/database-technologies/instant-client/overview/index.html).
I've added much more clear instructions in the Readme to reflect this. Let me know if there are other issues. Thanks!
