### 2019-01-14 mgckind/easyaccess#163 @ngoldbaum
Looks like that branch was merged. I looked over the new files and they look good to me. Closing.
