### 2019-01-14 mgckind/easyaccess#162 @ngoldbaum
> I think there is no issues here since we are nos distributing Oracle by any means

Not quite true, since you're running the `mgckind` conda channel which has binaries for the oracle instant client. It is true that you are allowed to do this, but must keep records when people download it as specified in the snippet of the license I linked to in November

Also I'd like to highlight the issue I originally pointed out: the oracle instant client is only licensed for testing and debugging purposes. If someone uses the instant client (or `easyaccess` itself) to do "production" work, they will be breaking the terms of the license and possibly exposing themselves to litigation from oracle.

So there are two issues here, only one of which is a user-facing problem:

* The oracle instant client license does not allow "production" work. Instead, for production work, people must use the full oracle client on e.g. a DES-managed system. This needs some sort of documentation in e.g. the README.

* The conda channel which easyaccess is distributed from may be breaking the oracle instant client license since users who download the client from there are not explicitly agreeing to the oracle license or having their contact information recorded as required by the redistribution clause of the oracle license.

I'm going to put my foot down on documenting this problem before we go further with the JOSS submission. It would also be very nice to get an expert opinion from e.g. someone at NUMFocus. It's possible that my interpretation of the license is incorrect, I'm not a lawyer after all, but I'd really like to get an expert opinion on that before dropping my objections in this issue.
