### 2019-01-14 #1022 @ngoldbaum
I've gone through the issues and I now only have one objection to publishing the paper: the license issues related to the oracle instant client.

I think at a minimum there needs to be a warning about using the oracle instant client for production work in the README. It would also be really nice if we could get an expert opinion from someone on the license issue.
