### 2019-01-14 mgckind/easyaccess#162 @ngoldbaum
Just in case there's a language issue or a misunderstanding, here are two more points that I think might help elucidate why I'm raising these issues:

• I agree that the oraclie client is “free to download, deploy, and distribute”, but that still means that you have to abide by the license. If Oracle has an issue with the instant client being made available on your conda channel will you be able to comply with their requests for records by giving details of all 196 people who have downloaded the client from your conda channel?
• The user may not be required to have the server installed, but if they use the instant client with `easyaccess` for “production” work (which could occur depending on how users use the data) they could be violating the instant client license. Even if it’s free to download, they may risk legal action.

I’m concerned about both of these issues and am not sure they have been addressed.

I agree that you’ve added clearer instructions in the README about how to download the instant client, but I don’t see adequate documentation of the terms of the instant client license or the requirement that the instant client only be used for testing and debugging purposes and *not* for production work.
