### 2019-01-14 mgckind/easyaccess#162 @mgckind
Thanks,
I think you are right, I'll get as much info as possible in order to move forward.
