### 2019-01-14 mgckind/easyaccess#162 @mgckind
For completeness only, anaconda had the client in their conda channel [here](https://anaconda.org/anaconda/oracle-instantclient), not the os-x version
