### 2019-01-14 mgckind/easyaccess#162 @ngoldbaum
Ah interesting, I guess a similar issue could be raised with anacona inc (assuming they don’t have a license waiver from Oracle).
