### 2019-01-14 mgckind/easyaccess#162 @mgckind
I'm contacting Oracle and asking around the building as well.
