### 2019-01-30 #1022 @mgckind
Thanks @arfon , this is what I got from Oracle representative close to the Oracle Instant client team

> You can only distribute a tool with Instant Client as long as the Instant Client terms are adhered to.  Since I am not a lawyer I should leave my comments at that; you should consult your own legal resources.
I'll wait a bit more and otherwise will remove the Oracle instant client from our conda account and rely on having it installed first, note that as I mentioned in an issue before the client is also distributed in the anaconda conda channel as well.
