### 2019-01-30 #1022 @arfon
OK, I've just heard back from the legal counsel at NumFOCUS:

> I looked at the terms of the license for cx_oracle and the way that this is a transitive dependency. I would say that 1) the troublesome terms have mostly disappeared for recent versions of the simple oracle client, and 2), I don't think that the Open Journals use of cx_oracle itself creates a "no commercial use" problem. To the extent that there is a problem, I think it would have to be with the Oracle DB itself, not the client library – but we are not responsible for that.

> So, based on my analysis, I would give this the green light.

✨
