### 2019-01-30 #1022 @arfon
Also @mgckind - @whedon has (experimental) abilities to suggest missing DOIs. The top one on this list might be a good suggestion? More generally, please make sure you've checked your bibtex entry for any missing DOIs for the references.
