+++ paper/paper.bib
@@ -128,14 +128,18 @@ archivePrefix = "arXiv",
 
 @ARTICLE{DR1,
    author = {{DES Collaboration}},
-    title = "{The Dark Energy Survey Data Release 1}",
-  journal = {ArXiv e-prints},
+       title = "{The Dark Energy Survey: Data Release 1}",
+  journal = {\apjs},
 archivePrefix = "arXiv",
    eprint = {1801.03181},
  primaryClass = "astro-ph.IM",
- keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies, Astrophysics - Solar and Stellar Astrophysics},
+ keywords = {astronomical databases: miscellaneous, catalogs, cosmology: observations, surveys, techniques: image processing, techniques: photometric },
      year = 2018,
-    month = jan,
-   adsurl = {http://adsabs.harvard.edu/abs/2018arXiv180103181A},
+    month = dec,
+   volume = 239,
+      eid = {18},
+    pages = {18},
+      doi = {10.3847/1538-4365/aae9f0},
+   adsurl = {http://adsabs.harvard.edu/abs/2018ApJS..239...18A},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
