### 2019-01-30 #1022 @mgckind
Thanks @arfon, and @ngoldbaum
I've updated one of the missing DOI, the other 2 papers do not have one.
Also, I pushed the code to Zenodo:
DOI : 10.5281/zenodo.2553673
