### 2019-01-30 #1022 @mgckind
Thanks @arfon . One more thing, is it possible to modify the xml so my last name appears as 'Carrasco Kind' instead of 'Kind' ?
https://github.com/openjournals/joss-papers/blob/master/joss.01022/10.21105.joss.01022.crossref.xml#L40
